requirements.txt
amqp==2.4.2
arrow==0.4.2
aspy.yaml==1.2.0
beautifulsoup4==4.7.1
billiard==3.6.0.0
bs4==0.0.1
celery==4.3.0
certifi==2019.3.9
cfgv==1.6.0
chardet==3.0.4
Django==2.2
django-celery-results==1.0.4
django-rest-framework==0.1.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
ics==0.4
identify==1.4.2
idna==2.8
importlib-metadata==0.9
kombu==4.5.0
nodeenv==1.3.3
pre-commit==1.15.2
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
toml==0.10.0
urllib3==1.24.2
vine==1.3.0
virtualenv==16.5.0
zipp==0.4.0
